requirements.txt
certifi==2020.6.20
chardet==3.0.4
facebook-sdk==3.1.0
idna==2.10
requests==2.24.0
selenium==3.141.0
urllib3==1.25.10
